+++ Resources/Prototypes/Entities/Structures/Storage/Closets/base_structureclosets.yml
@@ -98,6 +98,7 @@
     stateDoorClosed: generic_door
   - type: StaticPrice
     price: 50
+  - type: MesonsNonviewable
 
 # steel closet base (that can be constructed/deconstructed)
 - type: entity
