+++ Resources/Prototypes/Shaders/shaders.yml
@@ -129,8 +129,3 @@
   id: Scanline
   kind: source
   path: "/Textures/Shaders/scanline.swsl"
-
-- type: shader
-  id: BrightnessFilter
-  kind: source
-  path: "/Textures/Shaders/brightnessfilter.swsl"
